cheesboard.java
Since all chess boards available in the market are 8 X 8 boards, Alex decides to paint a customised N X N board. Given the painted chess board, can you tell if it is painted correctly or not ? A chess board is considered valid if every adjacent 2 cells are painted with different color. Two cells are considered adjacent if they share a boundary.